Refactored Unity process discovery into its own class without any MonoDevelop dependencies.
