+++ PlayerConnection.cs
@@ -30,8 +30,6 @@
 using System;
 using System.Collections.Generic;
 using System.Net.Sockets;
-using System.IO;
-using System.Runtime.InteropServices;
 using System.Net;
 using System.Net.NetworkInformation;
 using System.Text.RegularExpressions;
